requirements.txt
python-dotenv
groq
cohere
AppOpener
pywhatkit
beautifulsoup4
googlesearch-python
requests
rich
keyboard
mtranslate
selenium
webdriver-manager
pygame
edge-tts
pillow
psutil
pywin32
regex